requirements.txt
Django
django-photologue
psycopg2-binary
mysqlclient
ipython
